Corrected file name in main method
The file name in the main method of Assembler.java was changed from "input.txt" to "input". This was done to fix an error where the program was unable to locate the file using the previous name. The changes align the filename to the standard naming system used across various modules.